move to hexagon_common.cc
