+++ src/runtime/hexagon/hexagon/hexagon_common.cc
@@ -28,6 +28,7 @@
 #include "hexagon_common.h"
 
 #include <tvm/runtime/logging.h>
+#include <tvm/runtime/profiling.h>
 #include <tvm/runtime/registry.h>
 
 #include <sstream>
@@ -37,6 +38,10 @@
 
 #include "../../library_module.h"
 #include "hexagon_buffer.h"
+
+#if defined(__hexagon__)
+#include "HAP_perf.h"
+#endif
 
 namespace tvm {
 namespace runtime {
@@ -109,6 +114,28 @@ PackedFunc WrapPackedFunc(TVMBackendPackedCFunc faddr, const ObjectPtr<Object>& 
     }
   });
 }
+
+#if defined(__hexagon__)
+class HexagonTimerNode : public TimerNode {
+ public:
+  virtual void Start() { start = HAP_perf_get_time_us(); }
+  virtual void Stop() { end = HAP_perf_get_time_us(); }
+  virtual int64_t SyncAndGetElapsedNanos() { return (end - start) * 1e3; }
+  virtual ~HexagonTimerNode() {}
+
+  static constexpr const char* _type_key = "HexagonTimerNode";
+  TVM_DECLARE_FINAL_OBJECT_INFO(HexagonTimerNode, TimerNode);
+
+ private:
+  uint64_t start, end;
+};
+
+TVM_REGISTER_OBJECT_TYPE(HexagonTimerNode);
+
+TVM_REGISTER_GLOBAL("profiling.timer.hexagon").set_body_typed([](Device dev) {
+  return Timer(make_object<HexagonTimerNode>());
+});
+#endif
 }  // namespace hexagon
 
 namespace {
